Revert "Fix 1.12.1 authentication"
This reverts commit 47a28aab56e9b894667bde741af7aa465c03ccd9.

+++ src/Auth/RealmList.h
@@ -36,7 +36,7 @@ enum RealmFlags : uint8
  
 struct Realm
 {
-    uint32 Icon;
+    uint8 Icon;
     bool Lock;
     RealmFlags Flags;
     std::string Name;
